+++ packages/hmssdk_flutter/android/src/main/kotlin/live/hms/hmssdk_flutter/poll_extension/HMSPollQuestionExtension.kt
@@ -64,7 +64,7 @@ class HMSPollQuestionExtension {
                     return null
                 }
                 val duration = pollQuestion["duration"]?.let {
-                    it as Long
+                    (it as Int).toLong()
                 }?:run {
                     HMSErrorLogger.returnArgumentsError("duration should not be null")
                     return null
